feat: Bundle client code in server package
Since the ESM output of the client package cannot currently
run in Node.js in ESM mode, we may need to run end-to-end
tests in Node.js that interact with both the client and server code.

It makes sense to remove server code from the client bundle,
but not so much the other way around where size is less constrained.